+++ package.json
@@ -7,6 +7,7 @@
   "types": "target/index.d.ts",
   "scripts": {
     "task": "node --no-warnings= --loader ts-node/esm",
+    "serve": "docker run --rm --pull=always -it --env NETWORK_TYPE='single-node' --env PROOF_LEVEL='none' --env LOG_LEVEL='Info' -p 3085:3085 -p 5432:5432 -p 8080:8080 -p 8181:8181 -p 8282:8282 o1labs/mina-local-network:o1js-main-latest-lightnet",
     "test": "npm run task -- --test FungibleToken.test.ts",
     "build": "tsc"
   },
