+++ words.txt
@@ -5,6 +5,7 @@ esbenp
 harrysolovay
 kantp
 keypair
+lightnet
 merkle
 mina
 mintable
